requirements.txt
asttokens==3.0.0
boto3==1.37.3
botocore==1.37.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
copulas==0.12.1
ctgan==0.11.0
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
deepecho==0.7.0
exceptiongroup==1.2.2
executing==2.2.0
Faker==36.1.1
filelock==3.17.0
fonttools==4.56.0
fsspec==2025.2.0
graphviz==0.20.3
hiddenlayer==0.3
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lightning-utilities==0.12.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
onnx==1.17.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
prompt_toolkit==3.0.50
protobuf==6.30.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.1
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
rdt==1.14.0
requests==2.32.3
s3transfer==0.11.3
scikit-learn==1.6.1
scipy==1.15.2
sdmetrics==0.19.0
sdv==1.18.0
seaborn==0.13.2
six==1.17.0
sklearn-compat==0.1.3
soundfile==0.13.1
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.6.0
torchinfo==1.8.0
torchmetrics==1.6.1
torchsummary==1.5.1
torchviz==0.0.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
wfdb==4.2.0
